refactor: move unmarshallRecord to somewhere shareable
